Level up system completed

+++ Database/Database.sql
@@ -3,14 +3,14 @@ SET time_zone = "+00:00";
 
 CREATE TABLE `users`
 (
-    `serial` int(11) NOT NULL,
-    `guild_id` bigint(64) DEFAULT NULL,
-    `user_id` bigint(64) DEFAULT NULL,
-    `username` varchar(64) DEFAULT NULL,
-    `total_messages` int(11) DEFAULT 0,
-    `total_voice` int(11) DEFAULT 0,
-    `level` int(11) DEFAULT 0,
-    `xp` int(11) DEFAULT 0
+  `serial` int(11) NOT NULL,
+  `guild_id` bigint(64) DEFAULT NULL,
+  `user_id` bigint(64) DEFAULT NULL,
+  `username` varchar(64) DEFAULT NULL,
+  `total_messages` int(11) DEFAULT 0,
+  `total_voice` float DEFAULT 0.0,
+  `level` int(11) DEFAULT 0,
+  `xp` int(11) DEFAULT 0
 )ENGINE=InnoDB DEFAULT CHARSET=utf8mb3 COLLATE=utf8mb3_unicode_ci;
 
 ALTER TABLE `users`
